% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_wgb_url.R
\name{build_wgb_url}
\alias{build_wgb_url}
\title{Create the URLs for the pages containing the data on worldgovernmentbonds.com}
\usage{
build_wgb_url(country, metric, include_forecast)
}
\arguments{
\item{country}{character vector with the names of countries}

\item{metric}{character vector with the names of required metrics}

\item{include_forecast}{logical}
}
\value{
a tbl_df with columns country, metric and url as character classes
}
\description{
Create the URLs for the pages containing the data on worldgovernmentbonds.com
}
