% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{check_metric}
\alias{check_metric}
\title{check that the supplied metric is supported}
\usage{
check_metric(metric)
}
\arguments{
\item{metric}{character vector of metrics - options are 'rating', 'cds_5' and 'yield_' with a given maturity in years, e.g. 'yield_10'}
}
\value{
character vector of permissible metrics as a subset of supplied metrics
}
\description{
check that the supplied metric is supported
}
