% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{country_clean}
\alias{country_clean}
\title{standardise country names from 2 letter country iso}
\usage{
country_clean(country_iso)
}
\arguments{
\item{country_iso}{character vector containing iso2c}
}
\value{
character vector of full country names in lower case with hyphen separator
}
\description{
standardise country names from 2 letter country iso
}
