% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chart_data.R
\name{get_wgb_chart_data}
\alias{get_wgb_chart_data}
\title{get data from WGB charts}
\usage{
get_wgb_chart_data(pjs_session, metric)
}
\arguments{
\item{pjs_session}{remote driver object}

\item{metric}{string indicating the metric of interest}
}
\value{
df containing date, metric and value
}
\description{
needed for cds and yield data
}
