% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wgb_data.R
\name{get_wgb_data_list}
\alias{get_wgb_data_list}
\title{Get data from worldgovernmentbonds.com as a list}
\usage{
get_wgb_data_list(pjs_session, url_tbl)
}
\arguments{
\item{pjs_session}{phantom.js session}

\item{url_tbl}{object of class tbl_df containing columns url and metric}
}
\value{
a list of data.frames
}
\description{
Get data from worldgovernmentbonds.com as a list
}
