% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yld_forecast_data.R
\name{get_yld_forecast_data}
\alias{get_yld_forecast_data}
\title{get forecasts for yield data}
\usage{
get_yld_forecast_data(pjs_session, metric)
}
\arguments{
\item{pjs_session}{phantom.js session}

\item{metric}{string indicating the yield metric}
}
\value{
tbl_df with columns forecast_date of class \code{Date}, retrieve_date of class \code{Date}, metric of class \code{character} and value of class \code{numeric}
}
\description{
get forecasts for yield data
}
